BigInt, long long

+++ CppPractice/BigInt/BigInt.cpp
@@ -2,8 +2,8 @@
 
 void BigInt::Init()
 {
-	m_element = new unsigned int[MAXBIT];
-	for(int i = 0; i < MAXBIT; i++)
+	m_element = new unsigned long long[MAXBIT];
+	for(long long i = 0; i < MAXBIT; i++)
 		m_element[i] = 0;
 	m_isRefined = false;
 }
@@ -15,9 +15,9 @@ inline BigInt::BigInt()
 
 BigInt::BigInt(const BigInt& x)
 {
-	m_element = new unsigned int[MAXBIT];
+	m_element = new unsigned long long[MAXBIT];
 	m_isRefined = false;
-	for(int i = 0; i < MAXBIT; i++)
+	for(long long i = 0; i < MAXBIT; i++)
 		m_element[i] = x.GetElement(i);
 	Refine();
 }
@@ -33,8 +33,8 @@ BigInt::BigInt(char* n)
 		return;
 
 	Init();
-	int len = strlen(n);
-	int numBegin = 0;
+	long long len = strlen(n);
+	long long numBegin = 0;
 
 	if(*n == '-')
 	{
@@ -43,11 +43,11 @@ BigInt::BigInt(char* n)
 		numBegin = 1;
 	}
 
-	int index = 0;
-	for(int i = len - 1; i >= numBegin && index <= MAXBIT;)
+	long long index = 0;
+	for(long long i = len - 1; i >= numBegin && index <= MAXBIT;)
 	{
 		m_element[index] = 0;
-		for(int j = 1; j < JINZHI && i >= numBegin; j *= 10)
+		for(long long j = 1; j < JINZHI && i >= numBegin; j *= 10)
 		{
 			//if(n[i] == ' ' || n[i] == ',')
 			//	continue;
@@ -74,14 +74,14 @@ inline BigInt::~BigInt()
 	delete[] m_element;
 }
 
-int BigInt::GetElement(int i) const
+long long BigInt::GetElement(long long i) const
 {
 	if(i < 0 || i >= MAXBIT)
 		return -1;
 	return m_element[i];
 }
 
-void BigInt::SetElenemt(int i, int n)
+void BigInt::SetElenemt(long long i, long long n)
 {
 	if(i < 0 || i >= MAXBIT)
 		return;
@@ -95,8 +95,8 @@ void BigInt::Refine()
 {
 	if(m_isRefined)
 		return;
-	int carry = 0;
-	for(int i = 0; i < MAXBIT; i++)
+	long long carry = 0;
+	for(long long i = 0; i < MAXBIT; i++)
 	{
 		m_element[i] += carry;
 		if(m_element[i] >= JINZHI)
@@ -113,7 +113,7 @@ void BigInt::Refine()
 void BigInt::ToComplement()
 {
 	Refine();
-	for(int i = 0; i < MAXBIT - 1; i++)
+	for(long long i = 0; i < MAXBIT - 1; i++)
 		m_element[i] = JINZHI - 1 - m_element[i];
 	m_element[0] += 1;
 	Refine();
@@ -136,14 +136,14 @@ void BigInt::Print() const
 	}
 	else
 	{
-		int i = MAXBIT - 2;
+		long long i = MAXBIT - 2;
 		while(m_element[i] == 0 && i > 0)
 			i--;
 		std::cout << m_element[i] << " ";
 		i--;
 		for(; i >= 0; i--)
 		{
-			for(int tmp = JINZHI / 10; m_element[i] / tmp == 0 && tmp > 1; tmp /= 10)
+			for(long long tmp = JINZHI / 10; m_element[i] / tmp == 0 && tmp > 1; tmp /= 10)
 				std::cout << 0;
 			std::cout << m_element[i] << " ";
 		}
@@ -162,7 +162,7 @@ BigInt& BigInt::operator+(const BigInt& x) const
 {
 	BigInt* result = new BigInt();
 
-	for(int i = 0; i < MAXBIT; i++)
+	for(long long i = 0; i < MAXBIT; i++)
 		result->SetElenemt(i, GetElement(i) + x.GetElement(i));
 
 	result->Refine();
@@ -182,11 +182,11 @@ BigInt& BigInt::operator*(const BigInt& x) const
 	BigInt a = Abs();
 	BigInt b = x.Abs();
 
-	for(int i = 0; i < MAXBIT - 1; i++)
+	for(long long i = 0; i < MAXBIT - 1; i++)
 	{
 		if(x.GetElement(i) == 0)
 			continue;
-		for(int j = 0; j < MAXBIT - 1; j++)
+		for(long long j = 0; j < MAXBIT - 1; j++)
 		{
 			if(a.GetElement(j) == 0)
 				continue;
